Move Systray geometry methods into new auxiliary Systray::Geometry
This provides a slightly better separation of concerns.

The Qml position setting code is moved to C++, which allowed to get rid
of the currentScreenIndex() method (which seemed to be just a detour to
pass the screen from C++ to Qml).

Unfortunately, because Systray::Geometry is no Q_OBJECT, we lose the
Q_ENUM feature for TaskBarPosition.

Signed-off-by: Stephan Beyer <s-beyer@gmx.net>