Repair online/offline socket api commands
Our little dance from earlier to get to the journal db was leading to a
different convention for the relative path than what the syncmode table
was expecting (leading '/' vs no leading '/'). By relying on
FileData::get like everywhere else in SocketApi now we don't have that
problem anymore.

As a side effect we remove the now unused relativePath function in
Folder.

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>