Add "availability" column and enum to the sync logic
Intent is that it will replace the virtualfile column and the priority
list in the csync context. This will hopefully be a better way to model
the files lifecycle in the virtual drive context.

In the case of a regular sync the default is considering the file
available, while in the case of a virtual drive sync the default is
considering the file unavailable.

To guide the refactoring I added a bunch of asserts to make sure the
virtualfile and availability columns are always consistent.

This helped quite a lot in finding issues and also making sure we have
the right default values when we're syncing with or without the drive
logic enabled.

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>