Workaround the race condition on file open
When a drive implementation was requesting a sync, this would return
right away then the drive thread proceeding with handing out file
descriptors (for instance). So let's make the sync function a blocking
one instead, this way it ensures the needed files will be available
before we hand over the descriptors.

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>