Don't use the syncmode in the engine anymore
Modify the socket API implementation so that the availability enum is
changed when the sync mode is changed. For instance an unavailable file
that we request to be in offline mode will get the ItemNeedsDownload
flag.

This way, the syncengine itself needs to know only about the
availability enum it doesn't need to look at anything else.

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>